update: make paths and links more configurable but more coupled
