+++ v3/_posts/banff-trip.md
@@ -2,7 +2,7 @@
 title: 'Banff Trip'
 excerpt: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Praesent elementum facilisis leo vel fringilla est ullamcorper eget. At imperdiet dui accumsan sit amet nulla facilities morbi tempus.'
 coverImage: '/assets/images/banff_copy.png'
-date: '2024-12-05T05:35:07.322Z'
+date: '2020-03-16T05:35:07.322Z'
 author:
   name: Eri Lopez
   picture: '/assets/blog/authors/EriLopez3.png'
