+++ v3/components/hero-post.tsx
@@ -3,6 +3,7 @@ import DateFormatter from './date-formatter'
 import CoverImage from './cover-image'
 import Link from 'next/link'
 import type Author from '../interfaces/author'
+import { Sections } from '../lib/ConfigUtils'
 
 type Props = {
   title: string
@@ -10,7 +11,8 @@ type Props = {
   date: string
   excerpt: string
   author: Author
-  slug: string
+  slug: string,
+  currentTheme: Sections,
 }
 
 const HeroPost = ({
@@ -20,18 +22,19 @@ const HeroPost = ({
   excerpt,
   author,
   slug,
+  currentTheme
 }: Props) => {
   return (
     <section>
       <div className="mb-8 md:mb-16 w-full mx-auto h-[500px]">
-        <CoverImage title={title} src={coverImage} slug={slug} isHero />
+        <CoverImage title={title} src={coverImage} slug={slug} currentTheme={currentTheme} isHero />
       </div>
       <div className="md:grid md:grid-cols-2 md:gap-x-16 lg:gap-x-8 mb-20 md:mb-28">
         <div>
           <h3 className="mb-4 text-4xl lg:text-5xl leading-tight">
             <Link
-              as={`/posts/${slug}`}
-              href="/posts/[slug]"
+              as={`/${currentTheme}/${slug}`}
+              href={`/${currentTheme}/[slug]`}
               className="hover:underline"
             >
               {title}
