+++ v3/components/intro.tsx
@@ -1,9 +1,10 @@
+import { Sections } from "../lib/ConfigUtils"
 
-const Intro = () => {
+const Intro = ({ currentTheme }: {currentTheme: Sections}) => {
   return (
     <section className="flex-col md:flex-row flex items-center md:justify-between mt-16 mb-16 md:mb-12">
       <h1 className="text-5xl md:text-8xl font-bold tracking-tighter leading-tight md:pr-8">
-        Blog.
+        {currentTheme}
       </h1>
       <h4 className="text-center md:text-left text-lg mt-5 md:pl-8">
         I&apos;m Eri Lopez welcome to my blog 👋.
