+++ v3/components/post-header.tsx
@@ -3,23 +3,26 @@ import DateFormatter from './date-formatter'
 import CoverImage from './cover-image'
 import PostTitle from './post-title'
 import type Author from '../interfaces/author'
+import { Sections } from '../lib/ConfigUtils'
 
 type Props = {
   title: string
   coverImage: string
   date: string
   author: Author
+  currentTheme: Sections
 }
 
-const PostHeader = ({ title, coverImage, date, author }: Props) => {
+const PostHeader = ({ title, coverImage, date, author, currentTheme }: Props) => {
   return (
     <>
       <PostTitle>{title}</PostTitle>
       <div className="hidden md:block md:mb-12">
         <Avatar name={author.name} picture={author.picture} />
       </div>
+      {/* <div className="mb-8 md:mb-16 sm:mx-0"> */}
       <div className="mb-8 md:mb-16 sm:mx-0">
-        <CoverImage title={title} src={coverImage} />
+        <CoverImage title={title} src={coverImage} isHero={false} currentTheme={currentTheme}/>
       </div>
       <div className="max-w-2xl mx-auto">
         <div className="block md:hidden mb-6">
